WIP progress of GB Sounds engine
The GB Sounds engine has been ported and mostly works with the HQ mixer. I'm most likely missing some stuff as square and wave data aren't updating properly when swapping from a GB Sounds straight to m4a track and vice versa.

That aside, all songs from GSC are in the repo and one song from RBY has been ported to work with the engine (need to see if it sounds like it did in RBY)